Time: 6 ms (99.71%), Space: 47.9 MB (43.91%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -23,10 +23,10 @@ class Solution {
         return ans;
     }
 
-    int find(int mid,int[] piles){
-        int count=0;
+    double find(int mid,int[] piles){
+        double count=0;
         for(int i:piles){
-            count+=(double)(i+mid-1)/mid;
+            count+=(i+mid-1)/mid;
         }
         return count;
     }
